contact routing done WIP: search bar

+++ src/app/contacts/contact-detail/contact-detail.component.ts
@@ -50,10 +50,13 @@ export class ContactDetailComponent implements OnInit {
   }
 
   deleteContact() {
-    this.contactService.deleteContact(this.contact.contactID).subscribe();
-    setTimeout(() => {
+    if (window.confirm('Weet je zeker dat je dit contact wilt verwijderen?')) {
+      this.contactService.deleteContact(this.contact.contactID).subscribe();
+
+      // setTimeout(() => {
       this.router.navigate(['/contacten']);
-    }, 500);
+    }
+      // }, 500);
 
   }
